# backend/openapi.yaml
openapi: 3.1.0
info:
  title: MiroShark HTTP API
  version: "1.0.0"
  summary: Universal Swarm Intelligence Engine — REST surface for simulations, graphs, reports, sharing, and observability.
  description: |
    The MiroShark backend is a Flask service that orchestrates large agent
    populations through Wonderwall, builds bi-temporal knowledge graphs in
    Neo4j, and generates reports, share cards, and embeds for every run.

    This document is the canonical reference for the **HTTP** surface. The
    bundled MCP server (`backend/mcp_server.py`) exposes a parallel set of
    tools over stdio for Claude Desktop, Cursor, Windsurf and Continue —
    see `GET /api/mcp/status` and `docs/MCP.md` for that path.

    All responses are JSON unless explicitly marked otherwise (PNG share
    cards, file downloads, SSE streams). Successful responses follow the
    `{"success": true, "data": ...}` envelope used throughout the codebase;
    errors return `{"success": false, "error": "..."}` with an appropriate
    HTTP status.

    Endpoints are public on a default install — there is no auth layer in
    the bundled distribution. Operators are expected to bind the backend
    to a private interface or front it with a reverse proxy when deploying
    to production.

    **Mutation endpoints** that write to a simulation's on-disk state
    (`POST /publish`, `POST /resolve`, `POST /outcome`) are gated on a
    shared operator secret supplied via `Authorization: Bearer
...

...

  contact:
...
servers:
...
paths:
...
  /api/feed.atom:
    get:
      tags: [Publish & Embed]
      summary: Atom 1.0 syndication feed of the public simulation gallery.
      description: |
        Public MiroShark simulations rendered as an Atom 1.0 XML feed so
        researchers and tooling operators can subscribe via Feedly,
        Readwise, Inoreader, Obsidian RSS, NetNewsWire — every newly
        published simulation lands in their reader the same way an AI
        newsletter does.

        Mirrors `GET /api/simulation/public` selection (publish-gated,
        sorted by `created_at` desc, capped at 20 entries) so a
        subscription stays in lockstep with what shows up on `/explore`.
        Pure stdlib XML — zero new dependencies.

        Each entry carries: scenario as title, consensus split + quality
        + outcome label as summary, share-card PNG as `<media:thumbnail>`
        + `<media:content>`, replay GIF as `<media:content>` so River-
        view aggregators surface motion previews automatically.

        Browsers auto-discover this feed via the `<link rel="alternate"
        type="application/atom+xml">` tag in `index.html`.
      parameters:
        - in: query
          name: verified
          description: |
            Truthy values (`1`, `true`, `yes`, `on`) restrict the feed to
            simulations with a recorded outcome annotation — the same
            curated set the `/verified` page surfaces.
          schema: { type: string }
      responses:
        "200":
          description: |
            Atom 1.0 XML document with `Cache-Control: public, max-age=300`.
          content:
            application/atom+xml:
              schema: { type: string }
  /api/feed.rss:
    get:
      tags: [Publish & Embed]
      summary: RSS 2.0 syndication feed of the public simulation gallery.
      description: |
        Same content + selection as `/api/feed.atom` rendered as RSS 2.0
        for older readers and self-hosted aggregators that haven't moved
        off the legacy format. Published runs include their share-card
        PNG as an `<enclosure>` so podcast-style readers handle it
        natively.
      parameters:
        - in: query
          name: verified
          description: |
            Truthy values (`1`, `true`, `yes`, `on`) restrict the feed to
            simulations with a recorded outcome annotation.
          schema: { type: string }
      responses:
        "200":
          description: |
            RSS 2.0 XML document with `Cache-Control: public, max-age=300`.
          content:
            application/rss+xml:
              schema: { type: string }
...
components:
  securitySchemes:
    AdminToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Shared operator secret read from `MIROSHARK_ADMIN_TOKEN` in the
        backend process environment. Sent as `Authorization: Bearer
        <token>`. Compared in constant time on the server.

        Required on every mutation endpoint that writes to a
        simulation's on-disk state — `POST /publish`, `POST /resolve`,
        `POST /outcome`. Read endpoints (including `GET /outcome`)
        remain public.

        Fail-closed: when `MIROSHARK_ADMIN_TOKEN` is unset or empty in
        the deployment's environment, the gated endpoints return
        `503 — admin auth not configured` rather than allowing the
        request through. This is intentional — an operator who forgot
        to set the secret would otherwise ship an open mutation
        surface. Configure the token before exposing the backend
        publicly.
...